backend/requirements.txt
Flask==3.1.1
numpy==2.3.1
openai==1.93.0
pandas==2.3.0
pykrx==1.0.51
python-dotenv==1.1.1
